pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "linkedin-post-generator"
version = "0.1.0"
description = "AI-powered LinkedIn post generator with Telegram bot interface"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "python-telegram-bot[job-queue]>=21.0",
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",
    "feedparser>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
linkedin-post-gen = "main:main"
